+++ core/src/main/java/com/hortonworks/iotas/catalog/Topology.java
@@ -15,7 +15,7 @@ import java.util.Map;
  */
 public class Topology implements Storable {
 
-    public static final String NAME_SPACE = "topology";
+    public static final String NAME_SPACE = "topologies";
     public static final String ID = "id";
     public static final String NAME = "name";
     public static final String CONFIG = "config";