+++ core/src/main/java/com/hortonworks/iotas/storage/impl/jdbc/provider/phoenix/JdbcClient.java
@@ -1,9 +1,10 @@
-package com.hortonworks.iotas.storage.impl.jdbc.phoenix;
+package com.hortonworks.iotas.storage.impl.jdbc.provider.phoenix;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.BufferedReader;
+import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.sql.Connection;
 import java.sql.DriverManager;
@@ -11,23 +12,29 @@ import java.sql.DriverManager;
 /**
  * Basic PhoenixClient
  */
-public class PhoenixClient {
-    private static final Logger LOGGER = LoggerFactory.getLogger(PhoenixClient.class.getName());
+public class JdbcClient {
+    private static final Logger log = LoggerFactory.getLogger(JdbcClient.class.getName());
     private static final String DEFAULT_URL = "jdbc:phoenix:localhost:2181";
 
     private final String url;
 
-    public PhoenixClient() {
+    public JdbcClient() {
         this(DEFAULT_URL);
     }
 
-    public PhoenixClient(String url) {
+    public JdbcClient(String url) {
         this.url = url;
     }
 
     public void runScript(String path) throws Exception {
+        final InputStream resourceAsStream = this.getClass().getClassLoader().getResourceAsStream(path);
+        if(resourceAsStream == null) {
+            log.warn("Given resource with path [{}] could not be found", path);
+            throw new RuntimeException("Resource not found for given path "+path);
+        }
+
         try (BufferedReader reader =
-                     new BufferedReader(new InputStreamReader(this.getClass().getClassLoader().getResourceAsStream(path)));
+                     new BufferedReader(new InputStreamReader(resourceAsStream));
         ) {
             Connection connection = DriverManager.getConnection(url);
             connection.setAutoCommit(true);
@@ -36,7 +43,7 @@ public class PhoenixClient {
             while ((line = reader.readLine()) != null) {
                 if (!line.trim().isEmpty()) {
                     connection.createStatement().execute(line);
-                    LOGGER.info(String.format("######## SQL Query:  %s ", line));
+                    log.info(String.format("######## SQL Query:  %s ", line));
                 }
             }
         }
@@ -46,8 +53,8 @@ public class PhoenixClient {
         String url = "jdbc:phoenix:localhost:2181";
         String createPath = "phoenix/create_tables.sql";
         String dropPath = "phoenix/drop_tables.sql";
-        PhoenixClient phoenixClient = new PhoenixClient(url);
-        phoenixClient.runScript(dropPath);
-        phoenixClient.runScript(createPath);
+        JdbcClient jdbcClient = new JdbcClient(url);
+        jdbcClient.runScript(dropPath);
+        jdbcClient.runScript(createPath);
     }
 }