+++ core/src/main/resources/phoenix/drop_tables.sql
@@ -4,3 +4,9 @@ DROP TABLE IF EXISTS devices
 DROP TABLE IF EXISTS datasources
 DROP TABLE IF EXISTS topologies
 DROP TABLE IF EXISTS topology_components
+DROP TABLE IF EXISTS sequence_table
+DROP SEQUENCE IF EXISTS datasources_sequence
+DROP SEQUENCE IF EXISTS datafeeds_sequence
+DROP SEQUENCE IF EXISTS parser_info_sequence
+DROP SEQUENCE IF EXISTS topologies_sequence
+DROP SEQUENCE IF EXISTS topology_components_sequence