+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.1.2
+ * IOT-196: Phoenix storage provider support in webservices modules.
+ * IOT-231: Change device id and version to make and model.
  * IOT-215: Introduce new REST API to show metrics for topology.
  * IOT-170: Add file watcher and custom processor file watcher handler with unit tests.
  * IOT-224: Create tags on the fly for datasources for backward compatibility.